requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
passlib
python-jose
python-multipart